move to legacy: infer_types
