+++ weave/legacy/decorator_type.py
@@ -2,9 +2,9 @@ import dataclasses
 import inspect
 import typing
 
-from weave import errors, infer_types
+from weave import errors
 from weave import weave_types as types
-from weave.legacy import context_state, decorator_class, object_type_ref_util
+from weave.legacy import context_state, decorator_class, object_type_ref_util, infer_types
 
 _py_type = type
