+++ weave/legacy/op_def_type.py
@@ -11,9 +11,9 @@ import typing
 from _ast import AsyncFunctionDef, ExceptHandler
 from typing import Any
 
-from weave import environment, errors, infer_types, registry_mem, storage
+from weave import environment, errors, registry_mem, storage
 from weave import weave_types as types
-from weave.legacy import artifact_fs, artifact_local, context_state
+from weave.legacy import artifact_fs, artifact_local, context_state, infer_types
 
 if typing.TYPE_CHECKING:
     from .op_def import OpDef
